TASK 26

+++ app/Http/Controllers/SliderController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\Slider;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Facades\Validator;
 
 class SliderController extends Controller
 {
@@ -21,6 +22,16 @@ class SliderController extends Controller
 
     public function store(Request $request)
     {
+        $validator = Validator::make($request->all(), [
+            'title' => 'required|string',
+            'caption' => 'required|string',
+            'image' => 'required|image|mimes:jpg,jpeg,png|max:2048'
+        ]);
+
+        if ($validator->fails()) {
+            return redirect()->back()->withErrors($validator->errors())->withInput();
+        }
+
         $imageName = time().'.'.$request->image->extension();
         Storage::putFileAs('public/slider', $request->file('image'), $imageName);
 
@@ -43,6 +54,17 @@ class SliderController extends Controller
 
     public function update(Request $request, $id)
     {
+        $validator = Validator::make($request->all(), [
+            'title' => 'required|string',
+            'caption' => 'required|string',
+            'image' => 'image|mimes:jpg,jpeg,png|max:2048'
+        ]);
+
+        if ($validator->fails()) {
+            return redirect()->back()->withErrors($validator->errors())->withInput();
+        }
+
+
         if($request->hasFile('image')) {
             $oldImage = Slider::find($id)->image;
             Storage::delete('public/slider/' . $oldImage);
